+++ crates/egui_demo_lib/src/demo/demo_app_windows.rs
@@ -100,6 +100,7 @@ impl Default for DemoGroups {
                 Box::<super::tests::InputTest>::default(),
                 Box::<super::tests::LayoutTest>::default(),
                 Box::<super::tests::ManualLayoutTest>::default(),
+                Box::<super::tests::TessellationTest>::default(),
                 Box::<super::tests::WindowResizeTest>::default(),
             ]),
         }
@@ -365,13 +366,13 @@ mod tests {
     use crate::{demo::demo_app_windows::DemoGroups, Demo};
     use egui::Vec2;
     use egui_kittest::kittest::Queryable;
-    use egui_kittest::{Harness, SnapshotOptions};
+    use egui_kittest::{Harness, SnapshotOptions, SnapshotResults};
 
     #[test]
     fn demos_should_match_snapshot() {
         let demos = DemoGroups::default().demos;
 
-        let mut errors = Vec::new();
+        let mut results = SnapshotResults::new();
 
         for mut demo in demos.demos {
             // Widget Gallery needs to be customized (to set a specific date) and has its own test
@@ -405,12 +406,7 @@ mod tests {
                 options.threshold = 2.1;
             }
 
-            let result = harness.try_snapshot_options(&format!("demos/{name}"), &options);
-            if let Err(err) = result {
-                errors.push(err.to_string());
-            }
+            results.add(harness.try_snapshot_options(&format!("demos/{name}"), &options));
         }
-
-        assert!(errors.is_empty(), "Errors: {errors:#?}");
     }
 }