+++ crates/epaint/src/stroke.rs
@@ -69,21 +69,22 @@ pub enum StrokeKind {
     Outside,
 }
 
-impl Default for StrokeKind {
-    fn default() -> Self {
-        Self::Middle
-    }
-}
-
 /// Describes the width and color of paths. The color can either be solid or provided by a callback. For more information, see [`ColorMode`]
 ///
 /// The default stroke is the same as [`Stroke::NONE`].
-#[derive(Clone, Debug, Default, PartialEq)]
+#[derive(Clone, Debug, PartialEq)]
 #[cfg_attr(feature = "serde", derive(serde::Deserialize, serde::Serialize))]
 pub struct PathStroke {
     pub width: f32,
     pub color: ColorMode,
     pub kind: StrokeKind,
+}
+
+impl Default for PathStroke {
+    #[inline]
+    fn default() -> Self {
+        Self::NONE
+    }
 }
 
 impl PathStroke {
@@ -99,7 +100,7 @@ impl PathStroke {
         Self {
             width: width.into(),
             color: ColorMode::Solid(color.into()),
-            kind: StrokeKind::default(),
+            kind: StrokeKind::Middle,
         }
     }
 
@@ -114,11 +115,17 @@ impl PathStroke {
         Self {
             width: width.into(),
             color: ColorMode::UV(Arc::new(callback)),
-            kind: StrokeKind::default(),
+            kind: StrokeKind::Middle,
         }
     }
 
+    #[inline]
+    pub fn with_kind(self, kind: StrokeKind) -> Self {
+        Self { kind, ..self }
+    }
+
     /// Set the stroke to be painted right on the edge of the shape, half inside and half outside.
+    #[inline]
     pub fn middle(self) -> Self {
         Self {
             kind: StrokeKind::Middle,
@@ -127,6 +134,7 @@ impl PathStroke {
     }
 
     /// Set the stroke to be painted entirely outside of the shape
+    #[inline]
     pub fn outside(self) -> Self {
         Self {
             kind: StrokeKind::Outside,
@@ -135,6 +143,7 @@ impl PathStroke {
     }
 
     /// Set the stroke to be painted entirely inside of the shape
+    #[inline]
     pub fn inside(self) -> Self {
         Self {
             kind: StrokeKind::Inside,
@@ -168,7 +177,7 @@ impl From<Stroke> for PathStroke {
             Self {
                 width: value.width,
                 color: ColorMode::Solid(value.color),
-                kind: StrokeKind::default(),
+                kind: StrokeKind::Middle,
             }
         }
     }